Added test of nrml_to csv
